backend/requirements.txt
# --- Core API & Server ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9

# --- Real-Time (Webcam/Tracking) ---
websockets==13.1
JWT==1.4.4

# --- Security & Auth (JWT) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Database (PostgreSQL + pgvector) ---
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pgvector==0.3.6

# --- Machine Learning / Computer Vision ---
onnxruntime==1.19.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
